pyproject.toml
[project]
name = "fabric-to-espanso"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Hopsakee", email = "jdejong@posteo.nl"},
]
dependencies = [
    "ipykernel>=6.29.5",
    "markdown>=3.7",
    "pyyaml>=6.0.2",
    "qdrant-client>=1.12.1",
    "fastembed>=0.4.2",
    "streamlit>=1.41.1",
    "pyperclip>=1.9.0",
    "regex>=2024.11.6",
    "fastcore>=1.7.28",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
